pyproject.toml
[tool.poetry]
name = "library"
version = "0.1.0"
description = "API для управления библиотекой"
authors = ["Dmitrii Naumov <dmi3naumov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=5.2.7,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
djangorestframework-simplejwt = ">=5.5.1,<6.0.0"
drf-spectacular = ">=0.28.0,<0.29.0"
python-dotenv = ">=1.1.1,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
gunicorn = ">=23.0.0,<24.0.0"
whitenoise = ">=6.11.0,<7.0.0"
django-filter = "^25.2"
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
model-bakery = "^1.20.5"
black = "^25.9.0"
isort = "^6.1.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
